+++ bin/user/forecast_util.py
@@ -1,12 +1,27 @@
 from weewx.cheetahgenerator import SearchList
-import pprint
 
 # Copyright 2022 David Bätge
 # Distributed under the terms of the GNU Public License (GPLv3)
 
 
 class ForecastUtil(SearchList):
+    # TODO: obvis
+    # TODO: windGust & windChar is None
     def get_day_icon(self, summary):
-        pprint.pprint(summary)
-        #if summary.pop is not None:
-            #print(summary.pop)
+        day_icon = summary['clouds']
+
+        if summary['pop'].raw > 60:
+            if (summary['clouds'] == 'BK' or summary['clouds'] == 'B1' or
+                    summary['clouds'] == 'SC'):
+                if 'rain' in summary['precip']:
+                    day_icon = 'rain--scattered'
+                if 'snow' in summary['precip']:
+                    day_icon = 'snow--scattered'
+
+            if summary['clouds'] == 'B2' or summary['clouds'] == 'OV':
+                if 'rain' in summary['precip']:
+                    day_icon = 'rain'
+                if 'snow' in summary['precip']:
+                    day_icon = 'snow'
+
+        return day_icon + '.svg'