+++ install.py
@@ -107,6 +107,8 @@ class BasicInstaller(ExtensionInstaller):
                   'skins/weewx-wdc/includes/icons/forecast/SC.svg',
                   'skins/weewx-wdc/includes/icons/forecast/rain.svg',
                   'skins/weewx-wdc/includes/icons/forecast/snow.svg',
+                  'skins/weewx-wdc/includes/icons/forecast/snow--scattered.svg',
+                  'skins/weewx-wdc/includes/icons/forecast/rain--scattered.svg',
                   ]),
             ]
         )